fix layout and controller

+++ Modules/Perusahaan/database/migrations/2025_02_06_021722_create_shareholders_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             ->references('id')->on('akta_perusahaans')
             ->onDelete('cascade');
             $table->string('pemegang_saham'); 
-            $table->decimal('jumlah_saham', 15, 2); 
-            $table->decimal('saham_persen', 5, 2);
+            $table->integer('jumlah_saham'); 
+            $table->integer('saham_persen');
             $table->timestamps();
         });
     }
